requirements.txt
langchain-mcp-adapters
langchain-openai
fastapi
langchain-core
python-dotenv
langchain
uvicorn
langchain_huggingface
langchain_chroma
sentence-transformers
pydantic
langchain_google_genai
langchain_groq